improvements and refactoring
