+++ core/src/main/resources/application.conf
@@ -2,7 +2,7 @@
 # FlixDb Config File #
 ######################
 
-postgres-common {
+postgresql-common {
 
   user = ${PG_USER}
   user = "pguser"
@@ -17,25 +17,25 @@ postgres-common {
 }
 
 
-postgres {
+postgresql-main-pool {
 
-  user = ${postgres-common.user}
-  password = ${postgres-common.password}
-  host = ${postgres-common.host}
-  port = ${postgres-common.port}
-  database = ${postgres-common.database}
+  user = ${postgresql-common.user}
+  password = ${postgresql-common.password}
+  host = ${postgresql-common.host}
+  port = ${postgresql-common.port}
+  database = ${postgresql-common.database}
   maximumPoolSize = 20
   minimumIdle = 5
 
 }
 
-postgres-cdc {
+postgresql-cdc-pool {
 
-  user = ${postgres-common.user}
-  password = ${postgres-common.password}
-  host = ${postgres-common.host}
-  port = ${postgres-common.port}
-  database = ${postgres-common.database}
+  user = ${postgresql-common.user}
+  password = ${postgresql-common.password}
+  host = ${postgresql-common.host}
+  port = ${postgresql-common.port}
+  database = ${postgresql-common.database}
   maximumPoolSize = 2
   minimumIdle = 1
