--- core/src/test/resources/application.conf
@@ -1,11 +1,0 @@
-postgres-starved {
-
-  user = ${postgres-common.user}
-  password = ${postgres-common.password}
-  host = ${postgres-common.host}
-  port = ${postgres-common.port}
-  database = ${postgres-common.database}
-  maximumPoolSize = 1
-  minimumIdle = 1
-
-}
