+++ core/src/test/scala/com/flixdb/core/postgresql/SnapshottingSpec.scala
@@ -1,10 +1,11 @@
-package com.flixdb.core
+package com.flixdb.core.postgresql
 
 import java.util.UUID.randomUUID
 
 import akka.Done
 import akka.actor.ActorSystem
 import akka.testkit.TestKit
+import com.flixdb.core.EventEnvelope
 import com.typesafe.config.{ConfigFactory, ConfigParseOptions, ConfigSyntax}
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
@@ -77,6 +78,11 @@ abstract class SnapshottingSpec extends BaseSnapshottingSpec {
     snapshot = false
   )
 
+  val event6 = event1 // same as event1 but we'll write in a different namespace
+
+  val event7 = event2 // same as event2 but we'll write in a different namespace
+
+  val event8 = event3 // same as event3 but we'll write in a different namespace
 
   val snapshot: EventEnvelope = EventEnvelope(
     eventId = randomUUID().toString,
@@ -90,31 +96,40 @@ abstract class SnapshottingSpec extends BaseSnapshottingSpec {
     snapshot = true
   )
 
-  test("We can append some events") {
-    postgreSQL.createTablesIfNotExists("megacorp")
-    postgreSQL.appendEvents("megacorp", List(event1, event2, event3)).futureValue shouldBe Done
-    postgreSQL.appendEvents("megacorp", List(event4)).futureValue shouldBe Done
-    postgreSQL.appendEvents("megacorp", List(event5)).futureValue shouldBe Done
-    postgreSQL.selectEvents("megacorp", "accounts", "account-0").futureValue.size shouldBe 3
+  test("Appending some events") {
+    dataAccess.createTablesIfNotExists("megacorp")
+    dataAccess.createTablesIfNotExists("megacorp_backup")
+    dataAccess.appendEvents("megacorp", List(event1, event2, event3)).futureValue shouldBe Done
+    dataAccess.appendEvents("megacorp", List(event4)).futureValue shouldBe Done
+    dataAccess.appendEvents("megacorp", List(event5)).futureValue shouldBe Done
+    dataAccess.appendEvents("megacorp_backup", List(event6, event7, event8)).futureValue shouldBe Done
+    dataAccess.selectEvents("megacorp", "accounts", "account-0").futureValue.size shouldBe 3
   }
 
-  test("We can save a snapshot") {
-    postgreSQL.saveSnapshot("megacorp", snapshot).futureValue shouldBe Done
-    val selectEvents = postgreSQL.selectEvents("megacorp", "accounts", "account-0").futureValue
+  test("Taking a snapshot") {
+    dataAccess.snapshot("megacorp", snapshot).futureValue shouldBe Done
+    val selectEvents = dataAccess.selectEvents("megacorp", "accounts", "account-0").futureValue
     selectEvents.size shouldBe 1
     selectEvents.head shouldBe snapshot
   }
 
-  test("The snapshot that we took did not affect other streams or sub streams") {
+  test("Checking that the snapshot we took did not affect other streams or sub streams") {
 
-    val account1Events = postgreSQL.selectEvents("megacorp", "accounts", "account-1").futureValue
+    val account1Events = dataAccess.selectEvents("megacorp", "accounts", "account-1").futureValue
     account1Events.size shouldBe 1
     account1Events.head shouldBe event4
 
     val user0Events =
-      postgreSQL.selectEvents("megacorp", "users", "user-0").futureValue
+      dataAccess.selectEvents("megacorp", "users", "user-0").futureValue
     user0Events.size shouldBe 1
     user0Events.head shouldBe event5
+
+    val otherNamespaceEvents
+      = dataAccess.selectEvents("megacorp_backup", "accounts", "account-0").futureValue
+
+    otherNamespaceEvents.size shouldBe 3
+    otherNamespaceEvents shouldBe List(event6, event7, event8)
+
   }
 
 }
@@ -164,8 +179,8 @@ abstract class BaseSnapshottingSpec
     .parseString(
       s"""|container.host = "${postgreSQLContainer.getContainerIpAddress}"
           |container.port = ${postgreSQLContainer.getMappedPort(5432)}
-          |postgres.host = $${container.host}
-          |postgres.port = $${container.port}""".stripMargin,
+          |postgresql-main-pool.host = $${container.host}
+          |postgresql-main-pool.port = $${container.port}""".stripMargin,
       ConfigParseOptions.defaults().setSyntax(ConfigSyntax.CONF)
     )
     .resolve()
@@ -174,13 +189,13 @@ abstract class BaseSnapshottingSpec
 
   val mergedConfig = testConfig.withFallback(regularConfig)
 
-  val system = ActorSystem("flixdb", config = mergedConfig)
+  implicit val system = ActorSystem("flixdb", config = mergedConfig)
 
-  val postgreSQL: PostgreSQLExtensionImpl = PostgreSQL(system)
+  val dataAccess = new PostgresSQLDataAccessLayer()
 
   override def afterAll: Unit = {
     import scala.concurrent.duration._
-    postgreSQL.closePools()
+    dataAccess.closePools()
     TestKit.shutdownActorSystem(system, duration = 30.seconds, verifySystemShutdown = true)
     postgreSQLContainer.stop()
   }
